add repackage hilla-maven-plugin method to the script

+++ repackage-jars.sh
@@ -13,8 +13,7 @@ hilla
 vaadin-dev
 "
 
-# hilla-dev
-# artifacts=""
+#hilla-maven-plugin
 
 version="24.4.0.alpha2"
 local=~/.m2/repository/com/vaadin
@@ -24,14 +23,15 @@ ns=com.vaadin
 scr=`readlink -f "$0"`
 dir=`dirname $scr`/hack
 
-install() {
+mvnInstallFile() {
+  echo "Installing $1 in $L/$3/$version/$1"
   mvn -q install:install-file \
     -Dfile=$1 \
     -DpomFile=$2
 }
 
-for artifact in $artifacts; do
-  if expr $artifact : ".*hilla" > /dev/null; then
+computeVars() {
+  if expr $1 : ".*hilla" > /dev/null; then
     P=$base/hilla
     N=$ns.hilla
     L=$local/hilla
@@ -40,8 +40,34 @@ for artifact in $artifacts; do
     N=$ns
     L=$local
   fi
+}
 
+renameArtifact() {
+  echo "Renaming $1 to $2"
+  ns1=`echo $1 | cut -d : -f1`
+  art1=`echo $1 | cut -d : -f2`
+  ns2=`echo $2 | cut -d : -f1`
+  art2=`echo $2 | cut -d : -f2`
+
+  mv META-INF/maven/$ns1 META-INF/maven/$ns2
+  mv META-INF/maven/$ns2/$art1 META-INF/maven/$ns2/$art2
+
+  perl -pi -e 's|<groupId>'$ns1'</groupId>|<groupId>'$ns2'</groupId>|' META-INF/maven/plugin.xml META-INF/maven/$ns2/$art2/plugin-help.xml
+  perl -pi -e 's|<artifactId>'$art1'</artifactId>|<artifactId>'$art2'</artifactId>|' META-INF/maven/plugin.xml META-INF/maven/$ns2/$art2/plugin-help.xml
+  perl -pi -e 's|<goalPrefix>hilla</goalPrefix>|<goalPrefix>vaadin</goalPrefix>|' META-INF/maven/plugin.xml META-INF/maven/$ns2/$art2/plugin-help.xml
+  perl -pi -e 's|hilla:|vaadin:|g' META-INF/maven/plugin.xml META-INF/maven/$ns2/$art2/plugin-help.xml
+
+  perl -pi -e 's|artifactId='$art1'|artifactId='$art2'|' META-INF/maven/$ns2/$art2/pom.properties
+  perl -pi -e 's|groupId='$ns1'|groupId='$ns2'|' META-INF/maven/$ns2/$art2/pom.properties
+
+  perl -0777 -pi -e 's|(<groupId>)'$ns1'(</groupId>\s*<artifactId>)'$art1'(</artifactId>)|${1}'$ns2'${2}'$art2'${3}|ms' META-INF/maven/$ns2/$art2/pom.xml
+}
+
+
+for artifact in $artifacts; do
   cd $dir
+
+  computeVars $artifact
 
   echo "Repackaging $artifact $file $folder"
   folder=tmp_$artifact
@@ -55,27 +81,36 @@ for artifact in $artifacts; do
     if [ -f ../$artifact/$file ]; then
       cp ../$artifact/$file $file
     fi
-    install $file $file
+    mvnInstallFile $file $file $artifact
   else
     file=$artifact-$version.jar
-    curl -s -o ../$file $P/$artifact/$version/$file 
+    curl -s -o ../$file $P/$artifact/$version/$file
     jar -xf ../$file
-    for i in `find . -type f`; do
-      # echo "Processing $i ../$artifact/$i"
-      if [ -f ../$artifact/$i ]; then
-        c=`wc -c ../$artifact/$i | awk '{print $1}'`
-        if [ $c -eq 0 ]; then
-          echo "Removing $i"
-          rm $i
-        else
-          echo "Replacing $i"
-          rm $i
-          cp ../$artifact/$i $i
+    if [ $artifact = hilla-maven-plugin ]; then
+      old=$N:$artifact
+      N=vaadin.com
+      artifact=vaadin-maven-plugin
+      file=$artifact-$version.jar
+      renameArtifact $old $N:$artifact
+    else
+      for i in `find . -type f`; do
+        # echo "Processing $i ../$artifact/$i"
+        if [ -f ../$artifact/$i ]; then
+          c=`wc -c ../$artifact/$i | awk '{print $1}'`
+          if [ $c -eq 0 ]; then
+            echo "Removing $i"
+            rm $i
+          else
+            echo "Replacing $i"
+            rm $i
+            cp ../$artifact/$i $i
+          fi
         fi
-      fi
-    done
+      done
+      echo "Creating $file"
+    fi
+    echo "Creating $file"
     jar -cf ../$file .
-    echo "Installing $file in $L/$artifact/$version/$file"
-    install ../$file META-INF/maven/$N/$artifact/pom.xml
+    mvnInstallFile ../$file META-INF/maven/$N/$artifact/pom.xml $artifact
   fi
 done
